feature/FS06-set-mark-nxoai-TASK12 DONE - UPDATE CSS

+++ bktel/database/migrations/2023_11_13_033500_create_reports_table.php
@@ -19,12 +19,16 @@ class CreateReportsTable extends Migration
             $table->unsignedBigInteger('teacher_to_subject_id');
             $table->string('title');
             $table->string('path');
-            $table->integer('mark')->nullable();
+            
+            // Thay đổi cột mark thành kiểu số thực (float) và cho phép giá trị từ 0 đến 10
+            $table->float('mark', 8, 2)->nullable();
+            
             $table->string('note')->nullable();
             $table->timestamps();
-             // Thêm foreign key
-        $table->foreign('student_id')->references('id')->on('students');
-        $table->foreign('teacher_to_subject_id')->references('id')->on('teacher_to_subjects');
+
+            // Thêm foreign key
+            $table->foreign('student_id')->references('id')->on('students');
+            $table->foreign('teacher_to_subject_id')->references('id')->on('teacher_to_subjects');
         });
     }
